MainSentence를 FestivalSentenceItem으로 수정하고 FestivalSentenceItem을 문장 하나에 대한 것으로 수정했다
